Rename checkbox component
