+++ src/app/components/elements/inputs/StyledCheckbox.tsx
@@ -5,7 +5,7 @@ import {Spacer} from "../Spacer";
 
 export const StyledCheckbox = (props : InputProps) => {
     const [checked, setChecked] = useState(props.checked ?? false);
-    const id = useMemo(() => {return (props.id ?? "") + "-" + v4()}, [props.id]);
+    const id = useMemo(() => {return (props.id ?? "") + "-" + v4();}, [props.id]);
     return <div className="styled-checkbox-wrapper">
         <div className="mr-2 mb-3">
             {checked && <div className="tick"/>}
